Update: Move formatter service to client
